+++ src/components/layout/Tabs/styles.ts
@@ -1,0 +1,12 @@
+import { css } from '@emotion/core'
+
+import { C75_COLOR } from '../../../colors'
+
+export const defaultTabSyles = css`
+  padding: 5px;
+  background-color: ${`${C75_COLOR.toRGBA(0.2)}`};
+  border-radius: 3px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`